refactor: routerConfig, fix paths
