+++ src/entities/Comment/ui/CommentCard/CommentCard.tsx
@@ -1,6 +1,5 @@
 import { memo } from 'react';
 import { AppLink } from '../../../../shared/ui/appLink/AppLink';
-import { RoutePath } from '../../../../shared/config/routerConfig/routerConfig';
 import { Avatar } from '../../../../shared/ui/Avatar/Avatar';
 import { Text, TextSize } from '../../../../shared/ui/Text/Text';
 import { Skeleton } from '../../../../shared/ui/Skeleton/Skeleton';
@@ -8,6 +7,7 @@ import { classNames } from '../../../../shared/lib/classNames/classNames';
 import clss from './CommentCard.module.scss';
 import { Comment } from '../../model/types/comment';
 import { VStack } from '../../../../shared/ui/Stack';
+import { RoutePath } from '../../../../shared/const/router';
 
 interface CommentCardProps {
     className?: string;
