+++ src/shared/types/router.ts
@@ -1,0 +1,7 @@
+import { RouteProps } from 'react-router-dom';
+import { UserRole } from '../../entities/User';
+
+export type AppRoutesProps = RouteProps & {
+    authOnly?: boolean;
+    roles?: UserRole[];
+}
