+++ src/widgets/Sidebar/model/selectors/getSidebarItems.ts
@@ -1,6 +1,6 @@
 import { createSelector } from '@reduxjs/toolkit';
 import { getUserAuthData } from '../../../../entities/User';
-import { RoutePath } from '../../../../shared/config/routerConfig/routerConfig';
+import { RoutePath } from '../../../../shared/const/router';
 import MainIcon from '../../../../shared/assets/icons/main-20-20.svg';
 import AboutIcon from '../../../../shared/assets/icons/about-20-20.svg';
 import ProfileIcon from '../../../../shared/assets/icons/profile-20-20.svg';
